syscall: Avoid a compiler warning which turns into a bug. Simply cast the result to an int and life is good.

+++ src/sim/syscall_emul.cc
@@ -385,7 +385,7 @@ umaskFunc(SyscallDesc *desc, int num, LiveProcess *process, ThreadContext *tc)
     // changing anything.
     mode_t oldMask = umask(0);
     umask(oldMask);
-    return oldMask;
+    return (int)oldMask;
 }
 
 SyscallReturn
